Write to log file and mindim param added
Previously, you had to choose between writing to log file or to stdout. Made it difficult to track progress while still keeping record for future in log file. New wrapper function allows you to either
write to a log file only or pass verbose which writes to log file AND stdout.

Also added mindim parameter. It's convinient for restarting the run if it gets terminated mid-run. Also, good for skipping some of the low dimensional runs if crunched for time. Default is set to 20.